+++ changelog.md
@@ -29,4 +29,6 @@
  - updated latest users at 1st [user-serialized]
  - fixed edit profile of user role[fix/editing-userRole]
  - fixed UI of sidebar [fix/interface]
- - added show and hide password button[fix/button]
+ - added show and hide password button in login page[fix/button]
+ - added show and hide password button in user module[fix/user-passwordbtn]
+